requirements.txt
torch
torchvision
streamlit
scikit-learn
pillow
matplotlib
numpy
pandas